Made items Select more Precise
Updated the items select query so that it selects exactly what is needed. This is a good practice in general and protects against a new column being added later and then included on this return without being desired. This was one of the TODO's I recently added. Therefore, I removed that comment since it is done.